adding docker fiel for building + first changes to download captured content

+++ deploy/docker-build.sh
@@ -38,25 +38,26 @@ cd icu-4.8.1.1/source
 make && make install
 cd ..
 echo
+if [[ "$IS_IN_DOCKER_BUILD" != "yes" ]]
+then
+    echo "Recreating the build directory $BUILD_PATH..."
+    rm -rf $BUILD_PATH && mkdir -p $BUILD_PATH
+    echo
 
-echo "Recreating the build directory $BUILD_PATH..."
-rm -rf $BUILD_PATH && mkdir -p $BUILD_PATH
-echo
+    echo "Transferring the source: $SOURCE_PATH -> $BUILD_PATH. Please wait..."
+    cd $BUILD_PATH && cp -rp $SOURCE_PATH . && cd src
+    echo
 
-echo "Transferring the source: $SOURCE_PATH -> $BUILD_PATH. Please wait..."
-cd $BUILD_PATH && cp -rp $SOURCE_PATH . && cd src
-echo
+    echo "Compiling PhantomJS..." && sleep 1
+    python build.py --confirm --release --qt-config="-no-pkg-config" --git-clean-qtbase --git-clean-qtwebkit
+    echo
 
-echo "Compiling PhantomJS..." && sleep 1
-python build.py --confirm --release --qt-config="-no-pkg-config" --git-clean-qtbase --git-clean-qtwebkit
-echo
-
-echo "Stripping the executable..." && sleep 1
-ls -l bin/phantomjs
-strip bin/phantomjs
-echo "Copying the executable..." && sleep 1
-ls -l bin/phantomjs
-cp bin/phantomjs $SOURCE_PATH
-echo
-
+    echo "Stripping the executable..." && sleep 1
+    ls -l bin/phantomjs
+    strip bin/phantomjs
+    echo "Copying the executable..." && sleep 1
+    ls -l bin/phantomjs
+    cp bin/phantomjs $SOURCE_PATH
+    echo
+fi
 echo "Finished."
